% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_boxplot}
\alias{am_boxplot}
\title{am_boxplot object.}
\usage{
am_boxplot(x, ...)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid data}

\item{...}{ignored}
}
\description{
am_boxplot object.
}
\examples{
am_boxplot(example_grid)
am_boxplot(example_trouvelot)
}
