% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_dotplot}
\alias{am_dotplot}
\title{am_dotplot object.}
\usage{
am_dotplot(x, cbPalette, leg = "none", main = "Colonization", ...)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid data}

\item{cbPalette}{a vector of colors. Default is: c("#999999", "#E69F00", "#56B4E9",
"#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7").
    Colorbrewer and Colorgorical are website where is possible
    design nice color palettes.}

\item{leg}{this parameter allows to set the legend. Default is "none". The other
possible parameters are: "left", "right", "top", "bottom".}

\item{main}{Plot title. Default "Colonization".}

\item{...}{ignored}
}
\description{
am_dotplot object.
}
\examples{
am_dotplot(example_grid)
am_dotplot(example_trouvelot)
}
