% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_save}
\alias{am_save}
\title{am_save object.}
\usage{
am_save(x, ...)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid summary data}

\item{...}{ignored}
}
\description{
am_save object.
}
\examples{
am_save(am_summary(example_grid))
am_save(am_summary(example_trouvelot))
}
