% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_summary}
\alias{am_summary}
\title{am_summary object.}
\usage{
am_summary(x)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid data}
}
\description{
am_summary object.
}
\examples{
am_summary(example_grid)
am_summary(example_trouvelot)
}
