% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_grid}
\alias{example_grid}
\title{Grid system scoring table}
\format{A data frame with 63 rows and 7 variables:
\describe{
  It is mandatory that the dataset is formatted as follow
  \item{replicates}{Sample replicates}
  \item{samples}{Sample names}
  \item{Total}{Number of total fungal structures identified}
  \item{Hyphopodia}{Number of hypopodia identified}
  \item{IntrHyphae}{Number of interanl hyphae identified}
  \item{Arbuscule}{Number of arbuscules identified}
  \item{Vesicle}{Number of vesicles identified}
}}
\source{
PhD Caroline Gutjahr
}
\usage{
example_grid
}
\description{
A dataset containing the data for 4 samples: one control and three mutants.
The scoring system is the Grid one.
}
\keyword{datasets}
