% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_trouvelot}
\alias{example_trouvelot}
\title{Trouvelot system scoring table}
\format{A data frame with 1008 rows and 3 variables:
\describe{
  It is mandatory that the dataset is formatted as follow
  \item{scoring}{Scoring value}
  \item{replicates}{Sample replicates}
  \item{samples}{Sample names}
}}
\source{
Mara Novero
}
\usage{
example_trouvelot
}
\description{
A dataset containing the data for s34 samples: one control and two mutants.
The scoring system is the Trouvelot one.
}
\keyword{datasets}
