% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-grid.R
\name{gplot.grid}
\alias{gplot.grid}
\title{gplot object.}
\usage{
\method{gplot}{grid}(x, type = "boxplot", ...)
}
\arguments{
\item{x}{surveyor object}

\item{...}{ignored}
}
\description{
gplot object.
}
