% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read colonization files in.}
\format{A grid dataset should have 7 variables:
\describe{
  It is mandatory that the dataset is formatted as follow
  \item{replicates}{Sample replicates}
  \item{samples}{Sample names}
  \item{Total}{Number of total fungal structures identified}
  \item{Hyphopodia}{Number of hypopodia identified}
  \item{IntrHyphae}{Number of interanl hyphae identified}
  \item{Arbuscule}{Number of arbuscules identified}
  \item{Vesicles}{Number of vesicles identified}
}
A data frame with 1008 rows and 3 variables:
\describe{
  It is mandatory that the dataset is formatted as follow
  \item{scoring}{Scoring value}
  \item{replicates}{Sample replicates}
  \item{samples}{Sample names}
}

		 The order of the samples (in the "samples" column) is the order used for plot display. So,
		 if you like to have a specific order, please sort the original data accordingly}
\usage{
readData(infile)
}
\arguments{
\item{infile}{the name of the file which the data are to be read from.}
}
\value{
an object of class \code{grid} or \code{trouvelot}.
}
\description{
Reads data
}
\details{
This funtion reads trouvelot and grid data
}
\author{
Marco Chiapello <mc983@cam.ac.uk>
}
\keyword{IO,}
\keyword{file}
