% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_barplot}
\alias{am_barplot}
\title{am_barplot object.}
\usage{
am_barplot(x, cbPalette,  main = "Colonization",
                  stats = c("none", "asterisks", "letters"), ...)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid data}

\item{cbPalette}{a vector of colors. Default is: c("#999999", "#E69F00", "#56B4E9",
"#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7").
    Colorbrewer and Colorgorical are website where is possible
    design nice color palettes.}

\item{main}{Plot title. Default "Colonization".}

\item{stats}{Default is "none". If it is "asterisks" on the plot asterisks appear
below the sample statistically different from the control. The control
is the first sample in the input file. For statistical analysis check
`am_stat` function.}

\item{...}{ignored}
}
\description{
am_barplot object.
}
\examples{
am_barplot(example_grid)
am_barplot(example_trouvelot)
}
