% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-grid.R
\name{am_barplot}
\alias{am_barplot}
\title{am_barplot object.}
\usage{
am_barplot(x, ...)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid data}

\item{...}{ignored}
}
\description{
am_barplot object.
}
\examples{
am_barplot(example_grid)
am_barplot(example_trouvelot)
}
