% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_dotplot}
\alias{am_dotplot}
\title{am_dotplot object.}
\usage{
am_dotplot(x, cbPalette, stats = c("none", "asterisks", "letters"),
                   main = "Colonization",...)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid data}

\item{cbPalette}{a vector of colors. Default is: c("#999999", "#E69F00", "#56B4E9",
"#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7").
    Colorbrewer and Colorgorical are website where is possible
    design nice color palettes.}

\item{stats}{Default is "none". If it is "asterisks" on the plot asterisks appear
below the sample statistically different from the control. The control
is the first sample in the input file. For statistical analysis check
`am_stat` function.}

\item{main}{Plot title. Default "Colonization".}

\item{...}{ignored}
}
\description{
am_dotplot object.
}
\examples{
am_dotplot(example_grid)
am_dotplot(example_trouvelot)
}
