% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_dotplot}
\alias{am_dotplot}
\title{am_dotplot object.}
\usage{
am_dotplot(x, cbPalette,
				alpha = 0.05,
				annot = c("none", "asterisks", "letters"),
				method = c("none","holm","hommel", "hochberg",
						   "bonferroni", "BH", "BY", "fdr"),
         legend = c("right", "left", "top", "bottom"),
         main = "Gridline intersect method", ...)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid data}

\item{cbPalette}{a vector of colors. Default is: c("#999999", "#E69F00", "#56B4E9",
"#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7").
    Colorbrewer and Colorgorical are website where is possible
    design nice color palettes.}

\item{alpha}{Significant threshold}

\item{annot}{Default is "none". If it is "asterisks" on the plot asterisks appear
below the sample statistically different from the control. The control
is the first sample in the input file. If is is "letters" on the plot
letters appear below the samples groupping the samples based on the 
statistical test. For statistical tests check `am_stat` function.}

\item{method}{adjusts the p-value for multiple comparisons using the Bonferroni, Holm,
Hochberg, Bonferroni, Benjamini-Hochberg, Benjamini-Yekutieli or fdr 
adjustment
(see agricolae package for more details).
The default is no adjustment for multiple comparisons.}

\item{legend}{Indicate the legend position}

\item{main}{Plot title. Default "Colonization".}

\item{...}{ignored}
}
\description{
am_dotplot object.
}
\examples{
am_dotplot(example_grid)
am_dotplot(example_trouvelot)
}
