% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_save}
\alias{am_save}
\title{am_save object.}
\usage{
am_save(x, filename, ...)
}
\arguments{
\item{x}{summary, statistics or plot data from Trouvelot or Grid dataset}

\item{filename}{path where save the file}

\item{...}{arguments from ggsave function from ggplot2 package}
}
\description{
am_save object.
}
\details{
If you save the summary, two files will be saved corresponding to the
   summary two tables. The user can name the file but the save function
   will append "_per_Replicate.csv" and "_per_Sample.csv".

   If you save the statistical summary one file will be saved with the
   "_Stat.csv" extention.
}
\examples{
am_save(am_summary(example_grid))
am_save(am_summary(example_trouvelot))
am_save(am_barplot(example_trouvelot))
am_save(am_stat(example_trouvelot))
}
\seealso{
`ggsave` from ggplot2 package
}
