% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_save}
\alias{am_save}
\title{am_save object.}
\usage{
am_save(x, filename, ...)
}
\arguments{
\item{x}{summary, statistics or plot data from Trouvelot or Grid dataset}

\item{filename}{path where save the file}

\item{...}{ignored}
}
\description{
am_save object.
}
\examples{
am_save(am_summary(example_grid))
am_save(am_summary(example_trouvelot))
am_save(am_barplot(example_trouvelot))
}
\seealso{
`ggsave` from ggplot2 package
}
