% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_stat}
\alias{am_stat}
\title{am_stat object.}
\usage{
am_stat(x, method = c("none", "bonferroni", "sidak", "hs", "bh", "by"),
                             ...)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid data}

\item{method}{adjusts the p-value for multiple comparisons using the Bonferroni, Šidák, Holm,
Holm-Šidák, Hochberg, Benjamini-Hochberg, or Benjamini-Yekutieli adjustment
(see conover.test package for more details).
The default is no adjustment for multiple comparisons.}

\item{...}{ignored}
}
\description{
am_stat object.
}
\examples{
am_stat(example_grid)
am_stat(example_trouvelot)
}
