% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{am_stat}
\alias{am_stat}
\title{am_stat object.}
\usage{
am_stat(x, ...)
}
\arguments{
\item{x}{dataset containing Trouvelot or Grid data}

\item{...}{ignored}
}
\description{
am_stat object.
}
\examples{
am_stat(example_grid)
am_stat(example_trouvelot)
}
