% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read colonization files.}
\usage{
readData(infile)
}
\arguments{
\item{infile}{the name of the file which the data are to be read from.}
}
\value{
an object of class \code{grid} or \code{trouvelot}.
}
\description{
Reads data
}
\details{
This funtion reads trouvelot and grid data
}
\author{
Marco Chiapello <mc983@cam.ac.uk>
}
\keyword{IO,}
\keyword{file}
