% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read colonization files in.}
\format{
A grid dataset should have 7 variables:
\describe{
  It is mandatory that the dataset is formatted as follow
  \item{- Samples:}{Sample names}
  \item{- Replicates:}{Sample Replicates}
  \item{- Total:}{Number of total fungal structures identified}
  \item{- Hyphopodia:}{Number of hypopodia identified}
  \item{- IntrHyphae:}{Number of interanl hyphae identified}
  \item{- Arbuscules:}{Number of arbuscules identified}
  \item{- Vesicles:}{Number of vesicles identified}
}
A Trouvelot dataset should have 3 variables:
\describe{
  It is mandatory that the dataset is formatted as follow
  \item{- Samples:}{Sample names}
  \item{- Replicates:}{Sample Replicates}
  \item{- Scoring:}{Scoring value.
                    The only possible terms in this column are:
                    0A0,1A3,2A3,3A3,4A3,5A3,1A2,2A2,3A2,4A2,5A2,1A1,2A1,
                    3A1,4A1,5A1,1A0,2A0,3A0,4A0,5A0.
                    If 0 is present, it will be converted into 0A0}
}

         The order of the Samples (in the "Samples" column) is the order used for plot display. So,
         if you like to have a specific order, please sort the original data accordingly
}
\usage{
readData(infile, type = c("none", "trouvelot", "grid"))
}
\arguments{
\item{infile}{the name of the file which the data are to be read from.}

\item{type}{grid or trouvelot}
}
\value{
an object of class \code{grid} or \code{trouvelot}.
}
\description{
Reads data
}
\details{
This funtion reads trouvelot and grid data
}
\examples{
f <- dir(system.file("extdata", package = "Ramf"), full.names = TRUE, pattern = "grid.csv")
x <- readData(f, type = "grid")
}
\author{
Marco Chiapello <mc983@cam.ac.uk>
}
\keyword{IO,}
\keyword{file}
