% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read colonization files in.}
\usage{
readData(infile)
}
\arguments{
\item{infile}{the name of the file which the data are to be read from.}
}
\value{
an object of class \code{grid} or \code{trouvelot}.
}
\description{
Reads data
}
\details{
This funtion reads trouvelot and grid data

The imported datasets should be correctly formatted:
		 1. Grid dataset: 7 columns with the follow headers: "replicates", "samples",
			"Total", "Hyphopodia", "IntrHyphae","Arbuscule", "Vesicles"
		 2. Trouvelot datasets: 3 columns with the follow headers: "scoring", "replicates", "samples"

		 The order of the samples (in the "samples" column) is the order used for plot display. So,
		 if you like to have a specific order, please sort the original data accordingly
}
\author{
Marco Chiapello <mc983@cam.ac.uk>
}
\keyword{IO,}
\keyword{file}
